Redes/SistemaIoTDiscordBot/practica3-main/requirements.txt
discord.py
paho-mqtt
